# More efficient `vec_group_loc()` vector access