This PR further optimizes `vec_group_loc()`, which should directly result in performance improvements for `group_by()`.

In Instruments I noticed that the biggest killer in `vec_group_loc()` is the double deference in a tight loop of `INTEGER(VECTOR_ELT())`. That can be replaced with an array of integer pointers that index directly into those integer vectors. That really helped.

Side note: It is insane how fast this is if you order the data ahead of time. As pointed out by @brodieG

```r
library(dplyr, warn.conflicts = FALSE)
library(vctrs)

set.seed(42)

g <- sample(1e6, 1e7, replace = TRUE)
tbl <- tibble(g = g)

o <- order(g)
ordered_tbl <- tbl[o,]
```

```r
bench::mark(
  vec_group_loc(tbl),
  vec_group_loc(ordered_tbl),
  check = FALSE,
  iterations = 20
)

# Master
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 2 x 6
#>   expression                      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                 <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_group_loc(tbl)            2.86s    2.99s     0.333     163MB    0.416
#> 2 vec_group_loc(ordered_tbl) 277.95ms 304.13ms     3.18      163MB    3.33

# This PR
#> Warning: Some expressions had a GC in every iteration; so filtering is disabled.
#> # A tibble: 2 x 6
#>   expression                      min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>                 <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 vec_group_loc(tbl)            1.63s    1.71s     0.578     171MB    0.722
#> 2 vec_group_loc(ordered_tbl) 258.03ms 276.63ms     3.46      171MB    3.63
```