@brodieG 2020-03-12
Awesome!  Obviously the ratios are not quite as good once you include the ordering time, but at least for this 1e7/1e6 case it seems to be worth it.

If you do go ahead and end up using the order thing, please considering giving a nod to the data table folk that contributed the radix sort to R.  This trick is out of the question without that.

Also, as is probably obvious to you already, ordering might allow simplifying the group computation logic as well (maybe, I haven't looked at how you do it).

Thanks for keeping me in the loop, appreciate it.